Update main.py
This should take care of some of our problems
But there needs to be some debugging for if someone writes 'banana' or something instead of a number
And while the game over thing works now, it doesn't take into effect until another input. I have some ideas, but I haven't figured it out yet, so anyone would be welcome to try.